writing test for exchange
